% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt_content.R
\name{encrypt_content_html}
\alias{encrypt_content_html}
\title{Encrypt content for placing into html_document}
\usage{
encrypt_content_html(
  content,
  key,
  md = FALSE,
  mathjax = FALSE,
  character = TRUE,
  protect = TRUE
)
}
\arguments{
\item{content}{content to encrypt}

\item{key}{key to use for the encryption}

\item{md}{content is to be rendered as markdown on decryption}

\item{mathjax}{content is to be rendered as MathJax on decryption}

\item{character}{output as a character vector (if FALSE, an \code{htmltools} tag)}

\item{protect}{whether to wrap content in protection tags to keep pandoc from processing it}
}
\value{
Returns an protected HTML span containing the encrpyted information
}
\description{
Encrypt content for placing into html_document
}
