% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_setup.R
\name{encryptRmd_general_setup}
\alias{encryptRmd_general_setup}
\title{Perform encryptRmd setup for non-html_documents}
\usage{
encryptRmd_general_setup(hljs = FALSE)
}
\arguments{
\item{hljs}{Include highlight.js dependency? This will highlight R code after it is decrypted in source code.}
}
\value{
an htmltools tagList
}
\description{
This function should be called in the first chunk, with echo=FALSE and include=TRUE.
It will inject the necessary dependencies and setup knitr to transform the outputs and
plots from chunks.
}
