% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_document_enc.R
\name{html_document_enc}
\alias{html_document_enc}
\title{Output format to enable encrypted chunks in html_document}
\usage{
html_document_enc(...)
}
\arguments{
\item{...}{Arguments to be passed \code{rmarkdown::html_document}}
}
\value{
Returns an output_format that extends html_document
}
\description{
Output format to enable encrypted chunks in html_document
}
