% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_setup.R
\name{encryptRmd_general_setup}
\alias{encryptRmd_general_setup}
\title{Perform encryptRmd setup for non-html_documents}
\usage{
encryptRmd_general_setup(jquery = TRUE)
}
\arguments{
\item{jquery}{Include jquery? If the document type already includes it, this should be FALSE}
}
\value{
an htmltools tagList
}
\description{
This function should be called in the first chunk, with echo=FALSE and include=TRUE.
It will inject the necessary dependencies and setup knitr to transform the outputs and
plots from chunks.
}
